Fix parsing bug were ADTS header straddles buffers
The parser would previously have remembered that bytes for the rest of
the header were still required, but when the next buffer was delivered,
it would only read those remaining bytes from the source and then get
'stuck'.

We now loop when in this state so that we have the chance to reconsider
how much input to consume from the source buffer once we've successfully
read the whole header and know the frames length.