Remove lies about zero-copy from README
Actually, most of the data will be copied in the common case were an
ADTS frame is larger than an MPEGTS packet.

Should rework the API in future so that this really can be zrwo-copy, at
the cost of not delivering the entire ADTS frame in a single callback to
the calling code.